+++ lib/constants/app.dart
@@ -1,6 +1,6 @@
 /// BillPod - app-wide constants.
 ///
-// Time-stamp: <Wednesday 2026-04-29 15:15:37 +1000 Graham Williams>
+// Time-stamp: <Thursday 2026-05-28 14:10:23 +1000 Graham Williams>
 ///
 /// Copyright (C) 2026, Togaware Pty Ltd
 ///
@@ -36,11 +36,11 @@ const String appTitle = 'BilliPod - Manage Your Bills';
 /// info button). Supports basic Markdown.
 
 const String aboutText =
-    'BilliPod supports the management of your bills, upcoming, scheduled, '
-    'and past. A common workflow is to add a regular bill scheduled for the '
-    'next payment. When you have verified the payment has been successful '
-    'tap the **Duplicate** button to place the next bill in the '
-    '**Expected** list to be monitored.';
+    'BilliPod supports the management of your bills, whether that might be upcoming, scheduled, '
+    'or past bills. A common workflow is to add a regular bill scheduled for the '
+    'next payment. When you verify the payment within the app '
+    'tap the **Duplicate** button to place the next bill according to the **Frequency** in the '
+    '**Expected** list so that you can monitor it.';
 
 const String appDirectory = 'billipod';
 const String billsFileName = 'bills.ttl';